Simplify the babi_multiprocess.diff by capturing exceptions when obtaining babi models directy from the pool.
Capturing only on RPC requests was not enough because an RPC request to read
keywords, or attachments, for example, could lead to those models to try to
access the model which was not yet added to the pool.

That happened with keywords and attachments which have a reference field but
also trying to access the chart wizard if the server was reloaded while the
user had the tree view already loaded. Most probably other places suffered from
the same problem.